The solution for [NSETM-734](https://bbpteam.epfl.ch/project/issues/browse/NSETM-734).
This lib allows to apply Oren's algorithm of neuron reduction to Sonata
networks. Currently it does not have Unit tests, as it's API is under
the review too. Only essential integration tests are presented. The only
exposed API for now is `network_reduction.reduce_network` method. All
other stuff is additional. `SonataApi` is an adhoc wrapper around
`blubrain/snap`. It's purpose to represent the target Sonata network.
The lib is intended for Python3 only.

Change-Id: I7e79ad0394857d129d6a7e0c21fd3e7d80a0bb5c