Fix doc toctree length
This commit cleans up the length of toctree entries which made
it impossible to read the name of the section. It does this
by changing the autosummary files when the source is read.
They are now nested under a section called sonata_network_reduction
to make it clear they are submodules.

Additionally there are a few clean ups of docs (mainly newlines)
as well as .gitignore and build process (runs clean first).

Change-Id: Ib6125c1624d2aedbc5c126b7b58b245931833630